Postgres SQl

+++ routers/todos.py
@@ -29,8 +29,12 @@ class TodoRequest(BaseModel):
     complete: bool
 
 @router.get("/", status_code=status.HTTP_200_OK)
-async def read_all(db:Annotated[Session,Depends(get_db)]):
-    return db.query(Todos).all()
+async def read_all(user: user_dependency,  db:db_dependency):
+    if user is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
+                            detail="Authorization Failed")
+    return db.query(Todos).filter(Todos.owner_id == user.get('id')).all()
+
 
 @router.get("/todo/{todo_id}",status_code=status.HTTP_200_OK)
 async def read_todo(user: user_dependency, db: db_dependency,
@@ -38,7 +42,8 @@ async def read_todo(user: user_dependency, db: db_dependency,
     if user is None:
         raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED,
                             detail="Authorization Failed")
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+           .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is not None:
         return todo_model
     raise HTTPException(status_code=404, detail="Todo not found")
@@ -65,10 +70,18 @@ async def create_todo(
 
 
 @router.put("/todo/todo_id", status_code=status.HTTP_200_OK)
-async def update_todo(db: db_dependency,
+async def update_todo(user:user_dependency, db: db_dependency,
                       todo_id: int,
                       todo_request: TodoRequest):
-    todo_model = db.query(Todos).filter(Todos.id == todo_id).first()
+    if user is None:
+        raise HTTPException(
+            status_code=401,
+            detail="Authorization Failed"
+        )
+
+
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+            .filter(Todos.owner_id == user.get('id')).first()
     if todo_model is None:
         raise HTTPException(status_code=404,deial="Todo not found")
 
@@ -80,14 +93,20 @@ async def update_todo(db: db_dependency,
     db.add(todo_model)
     db.commit()
 
-@router.delete("/todo/{todo_id}",status_code=status.HTTP_204_NO_CONTENT)
-async def delete_todo(db: db_dependency, todo_id: int = Path(gt=0)):
-                      todo_model = db.query(Todos).filter(Todos.id == todo.id).first()
-                      if todo_model is None:
-                        raise HTTPException(status_code=404, detail="Todo not found")
-                      db.query(Todos).filter(Todos.id == todo_id).delete()
+@router.delete("/todo/{todo_id}", status_code=status.HTTP_204_NO_CONTENT)
+async def delete_todo(user: user_dependency, db: db_dependency, todo_id: int = Path(gt=0)):
+    if user is None:
+        raise HTTPException(
+            status_code=401,
+            detail="Authorization Failed"
+        )
 
-                      db.commit()
+    todo_model = db.query(Todos).filter(Todos.id == todo_id)\
+        .filter(Todos.owner_id == user.get('id')).first()
 
+    if todo_model is None:
+        raise HTTPException(status_code=404, detail="Todo not found")
 
+    db.query(Todos).filter(Todos.id == todo_id).filter(Todos.owner_id == user.get('id')).delete()
+    db.commit()
